+++ metadata.rb
@@ -4,16 +4,14 @@ maintainer_email 'cookbooks@chef.io'
 license 'Apache-2.0'
 description 'Sets up the Docker yum repository'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version '2.1.0'
+version '2.2.0'
 
 recipe 'chef-yum-docker::default', 'Sets up the Docker yum repository.'
 
-%w(centos redhat scientific oracle amazon fedora).each do |os|
+%w(centos redhat scientific fedora).each do |os|
   supports os
 end
 
-depends 'compat_resource', '>= 12.16.3'
-
 source_url 'https://github.com/chef-cookbooks/chef-yum-docker'
 issues_url 'https://github.com/chef-cookbooks/chef-yum-docker/issues'
-chef_version '>= 12.1'
+chef_version '>= 12.5' if respond_to?(:chef_version)